# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CH RTV Platform API",
    "version": "1.0.9",
    "description": "Admin API and platform entry points for the Carrier Haulage Real-Time Visibility platform."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "admin-api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/orders/{id}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Update transport order",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "Delete transport order",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted"
          }
        }
      }
    },
...
    "/api/facilities/{id}": {
      "put": {
        "tags": [
          "Facilities"
        ],
        "summary": "Update facility",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Facility updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Facilities"
        ],
        "summary": "Delete facility",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Facility deleted"
          }
        }
      }
    },
...
    "/api/assignments/{id}": {
      "put": {
        "tags": [
          "Assignments"
        ],
        "summary": "Update assignment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment updated"
          }
        }
      }
    },
...
  }
}